modules/backend/src/vaultoidc/test_requirements.txt
moto==2.0.6
pytest==6.2.2
pytest-mock==3.6.1
